Excel Workbook
This workbook has the dataset towards the left side of the sheet and on the same sheet it also has pivot tables which summarizes the Production value in every state, number of stocks in each state and also yearly production value along with the grand totals. When we click on any value/cell, in the pivot table, to understand how this value is generated/calculated, we get sheets to the left of the current sheet which shows the corresponding tables/data that resulted the particular value. The sheets "Alabama" and "South Carolina" are the examples that shows the rows which were used to calculate its production values.